Add basic GUI
Four buttons (to select which sorting algorithm you would like to
perform), and two labels to show the list of numbers you are sorting
before and after having sorted them.

Currently all of the command line options are just commented out. This
will change in future versions of the code.