Make GUI / Command line versions mutually exclusive
The user can either run the application with the -g/--gui flag, in which
case a GUI will be created. Otherwise, the user can specify which
sorting algorithm to use via the appropriate flags.

The two options are mutually exclusive (either GUI or command line
sorting algorithm), but at least one MUST be specified